Queries Controllers 🍌
I redid all the queries of the api. Now I make three queries.

One to fetch all threads. Another to get all the questions related to a specific thread. Last but not least, I do a third query to get the data related to a specific question, including comments on that question.