Send Email Service 📮
I created a function that receives user input as parameters and through that same data hydrates a simple email message template and then sends that same message to the indicated user.

Then I just had to add this same function in the Sign Up controller, as soon as the user is inserted into the database, we can then send the email with the account data.