Threads UI - Changed 🎃
I made some changes to the UI of the threads and the UI of the question. However, I forgot to add the page where all the questions related to a thread would be presented.

I twisted the illustrations of the backgrounds and even made some small adjustments on how the threads, questions and also just one question would be displayed.