Component State & Testing 🍏
I added useState to the sign up and sign in page, as well as making changes to the application tests.

And to make everything much faster, I added a function that randomizes a string, which was used to randomize the usernames, emails and passwords used in the application tests.

Speaking of tests, I removed the existing tests and changed the methodology. Instead of doing for features, I do it for usage scenarios. Such as the workflow of a new user on the platform.