Threads and Questions UI - almost done 🍟
I made changes to the UI of the threads and questions page. In these I created some components to be able to pass props so that it is more dynamic and simpler to maintain their code.

In addition to this I changed a line in the query of the questions in the backend, added the name of the thread in the response object so that in the frontend I can put in the tag of the questions component the thread to which it belongs.