Modals UI 🍼
I created several modals related to a single question. That is, I created modals to edit, delete and comment on a specific issue. In the same way that I created the modals to edit and delete a specific comment. All data is dynamic and the components are in use with react's useState. Ready to be consumed.