frontend auth 🍬
I already created the instance of axios to have the base endpoint to communicate with the api.

In the same way that I created the handlers for Sign Up, Sign In and Sign Out. Just as I ended up updating the Cypress test for a new user, I ended up finding that everything works as it should, from frontend to backend.

In addition to this, I have already created the basic zustand store to store data related to the user, such as username, id and token.